+++ cmd/cli/test/armaria_test.go
