+++ cmd/cli/test/features/cli_add_tag.feature
