+++ cmd/cli/test/features/cli_query.feature
