+++ cmd/cli/test/features/cli_remove_book.feature
