+++ cmd/cli/test/features/cli_remove_folder.feature
