+++ cmd/cli/test/features/cli_remove_tag.feature
