+++ cmd/cli/test/test_helpers.go
@@ -12,7 +12,7 @@ import (
 	"github.com/cucumber/godog"
 	"github.com/google/shlex"
 	"github.com/google/uuid"
-	"github.com/jonathanhope/armaria/cmd/cli/cmd"
+	"github.com/jonathanhope/armaria/cmd/cli/internal"
 	"github.com/jonathanhope/armaria/internal/null"
 	"github.com/jonathanhope/armaria/pkg/api"
 	"github.com/jonathanhope/armaria/pkg/model"
